backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
Pillow>=10.0.0
torch>=2.0.0
